backend/src/api/auth_middleware.py
import jwt
from fastapi import Request, HTTPException, Security
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from src.database import settings

security = HTTPBearer()

def verify_jwt(auth: HTTPAuthorizationCredentials = Security(security)):
    try:
        payload = jwt.decode(
            auth.credentials,
            settings.auth_jwt_secret,
            algorithms=["HS256"]
        )
        user_id: str = payload.get("sub")
        if user_id is None:
            raise HTTPException(status_code=401, detail="Invalid token")
        return int(user_id)
    except jwt.ExpiredSignatureError:
        raise HTTPException(status_code=401, detail="Token expired")
    except jwt.InvalidTokenError:
        raise HTTPException(status_code=401, detail="Invalid token")
