backend/requirements.txt
# Backend requirements
fastapi
sqlmodel
uvicorn
pyjwt
pydantic-settings
psycopg2-binary
pytest
